change player eh to percentages and attacks cost e

+++ Source/Game/UI/eh.h
@@ -4,13 +4,13 @@
 #include "../Bittermap.h"
 
 class EH {
-    int const  *_health;
-    int const  *_energy;
+    float const  *_health;
+    float const  *_energy;
 public:
     void Init();
     void Show();
-    void setHealth(int const  *health) {_health = health;}
-    void setEnergy(int const  *energy) {_energy = energy;}
+    void setHealth(float const  *health) {_health = health;}
+    void setEnergy(float const  *energy) {_energy = energy;}
     game_framework::Bittermap _frame;
     game_framework::Bittermap _innerBar;
 };
